Refactor AutomapperBootstraper
